+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Add truncation to `VoteProposalDateListItem` module component.
+
 ## [1.0.47] - 2024-09-23
 
 ### Added