+++ packages/gov-ui-kit/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Bump `actions/setup-node` from 4.0.3 to 4.0.4
 -   Update minor and patch NPM dependencies
+-   Bump `@rollup/plugin-typescript` from v11 to v12 and `@rollup/plugin-commonjs` from v26 to v28
 
 ## [1.0.47] - 2024-09-23