+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Bump `actions/checkout` from 4.2.0 to 4.2.1
+
 ## [1.0.48] - 2024-10-04
 
 ### Added