+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 -   Bump `actions/checkout` from 4.2.0 to 4.2.1
+-   Update minor and patch NPM dependencies
 
 ## [1.0.48] - 2024-10-04